`test_expand` operational
