+++ ldtest/data/subclasses.yaml
@@ -2,5 +2,4 @@ $id: python://ldtest.JSONLDTests
 iolanta:supports:
   $id: pytest
 iolanta:isFacetOf:
-  - $id: tests:PositiveEvaluationTest
-  - $id: tests:NegativeEvaluationTest
+  - $id: tests:ExpandTest
